Add dedicated functions for virt-chroot execution
Important values like chroot binary path, default mount namespace
and mount execution options are not defined in a single place.
Therefore, the execution of virt-chroot, especially mounting/unmounting,
is duplicated and spread arround the code

Now dedicated methods - MountChRoot(), UnmountChRoot() and ExecChroot()
- are added to virt-chroot.go file under util package. In addition,
GetChrootBinaryPath() and GetChrootMountNamespace() are added,
returning the data which is saved in a single place inside that file.

Signed-off-by: Itamar Holder <iholder@redhat.com>