network: move the cache methods to the podNIC implementations
This commit moves the load / set cache methods from the
`BindMechanism` implementations to the `podNIC` implementations.

The VIF caches must be handled at a later stage, since it does not
use the new cache mechanism introduced in [0]; this caching scheme
will be very helpful to allow the VIF caches to be properly unit
tested.

[0] - https://github.com/kubevirt/kubevirt/pull/5054

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>