tests, networking, slirp: split outside conn tests
intrnal and outside connectivity tests are two different logical parts, when
testing network performance, and therefore should be tested separately.
One use-case for example, that will be addressed in the following commit, is
to be able run the outside checks without constantly failing, on an environment
without internet connectivity. splitting the tests will make the outside conn test
much easier to skip.

This commit extracts the outside connectivity checks in to different test cases.

Signed-off-by: alonsadan <asadan@redhat.com>