network, bind mech: add generateDomainIface method
This new method takes up some responsability from the
`preparePodInterface` method, which does too much (creates /
configures networking infrastructure, mutates the struct members,
and returns the domain interface).

Concretelly the domain interface generation is moved to this new
method.

The tap device name is added to the bridge binding mechanism struct
so that the created tap device name matches the domain interface
`Target` name - it wouldn't otherwise, since the
`generateTapDeviceName` returns different names depending on when
it is called (before switching the pod interface with a dummy).

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>