Consolidate rules added to output chain
This commit adds destination addresses to match for dnat
in a single rule.

By default traffic destined to 127.0.0.1 is DNATed.

With istio, we need to additionally take care of
passthrough traffic (traffic destined to ports not declared in a k8s)
which is sent from 127.0.0.6 to pod IP.
Since we're using destination address to match traffic
for dnatting, this commit reads pod IP of the virt launcher
and adds it to the set of matched IP addresses, if istio is used.

Signed-off-by: Radim Hrazdil <rhrazdil@redhat.com>