Allow kubevirt-operator to discover kube-prometheus
If no monitoring namespace is configured, check in addition to the
`openshift-monitoring` namespace also for the `monitoring` namespace.

https://github.com/prometheus-operator/kube-prometheus deploys
prometheus-operator in a similar enough fasion to openshift that simply
checking if the `monitoring` namespace exists in addition to
`openshift-monitoring` is enough to support autodiscover of
kube-prometheus and automatically deploy our monitoring objects on the
cluster. The same serviceAccount name is used.

`openshift-monitoring` has precedence for backward compatibility.

Signed-off-by: Roman Mohr <rmohr@redhat.com>