network, bridge: do not use dhcpConfig object on preparePodNetwork
Use members on the bridge binding mechanism instead, or directly
from the podNIClink, which is already a member.

Follow-up commits will introduce a new method to the bind
mechanism interface to ensure that each of its methods does exactly
a single thing: cache the pod link, and generate the dhcp config.

This separation will prepare the ground for the dhcpConfig to be
dropped as member of the bind mechanism implementations that
require DHCP.

Since we now read the MTU directly from the podniclink struct, we
need to update the MTU of the primary interface after its name was
changed.

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>