Enable live migration of VMs with hotplug volumes
The logic is implemented as follows:

- For the migration with the status MigrationPending after creating the
virt-launcher target pod, virt-controller checks if the VMI has hotplug
volumes attached. In case if it does, a new target attachment pod is
created (with the target virt-launcher pod set as owner).

- When the attachment pod is ready virt-controller updates the migration
phase to MigrationScheduled.

- Once in MigrationScheduled, virt-controller adds the target attachment
pod UID to the vmi.Status.MigrationState struct (the UID is needed for
the virt-handler on the target node).

- During the preparation stage virt-handler on the target node takes the
attachment pod UID and performs the bind mount of the hotplug volumes to
the new virt-launcher pod.

Signed-off-by: Vasiliy Ulyanov <vulyanov@suse.de>