virt-launcher: fix UEFI-mode without secureboot
This commit fixes UEFI mode with secureboot disabled
for base images (for virt-launcher) where only `OVMF_CODE.secboot.fd`
is shipped, letting the combination `OVMF_CODE.secboot.fd` +
`OVMF_VARS.fd` to be used to correctly boot in EFI mode with
no secureboot.

Previously when disabling secureboot virt-launcher would only try
to use `OVMF_CODE.fd` + `OVMF_VARS.fd` resulting in a VMI failure
when the former firmware file was not shipped.

Unfortunately a couple unit test had to be removed as the generated
VMI XML domain is now dependent on which files are present in the
filesystem and the FS cannot be mocked.

Signed-off-by: Antonio Cardace <acardace@redhat.com>