Move duplication check to CI only
This is slightly inconvenient, but locally in `bazel-testlogs` there are
sometimes stale entries from moved packages which make the check fail.
It is inconveninent to be not informed locally, but at least on CI it is
right now 100% correct.

This should be a temporary measure. We can check for non-existing
packages in `bazel-buildlogs` and remove them on the long run.

Signed-off-by: Roman Mohr <rmohr@redhat.com>