When comparing image sizes being equal, tolerate sub-512 byte difference
Our measurement of image size for the OCS case is very precise and
causes unchanged images to appear to change their size by a few bytes.
Tolerating very small changes reduces the flakiness of this test.

Such small differences are equal from the perspective of a VM, which
wouldn't make accesses under 512 bytes anyway, as that is the smallest
hardware block size.

Signed-off-by: Maya Rashish <mrashish@redhat.com>