hack: fix root owned _out directory issue
There is an issue with _out directory being root-owned
when building the Kubevirt project for the first time on
a clean environment.

The root cause is that we don't explicitly create the
_out directory, leaving it to be done by rsync during the
sync out procedure. Therefore rsync creates the _out directory
with incorrect ownership, thus _out become root-owned.

Proposed solution is just to explicitly create the OUT_DIR
in case SYNC_OUT is turned on.

Signed-off-by: Igor Bezukh <ibezukh@redhat.com>