fix test by changing it to run without block mode
The test keeps failing due to different problems
with a print "resized successfully" beeing printed
unexpectedly (which is due the fact that the test is with block mode).
The origin fix was trying to expect for that print -
make sure the action occured and only then continue
to the actual test, this doesnt seem to work by weird
logging behavior in the CI only.
The block mode was added because previously the default PVC fs was xfs,
Now the default fs is ext4 so no need for the block mode
which causes the problematic print.

Signed-off-by: Shelly Kagan <skagan@redhat.com>