Assign correct memory sizes to numanodes for hugepages
If the memory is divisable by the requested hugepagesize, it may still
not be equally distributable between the nodes. To overcome this, first
split the requested memory if necessary in smaller equally sized chunks
and then assign the rest of the numa nodes one by one to the numa nodes.

Signed-off-by: Roman Mohr <rmohr@redhat.com>