Detect driver cache mode for cloud-init disk
The cloud-init ISO is not yet generated at the time when driver cache
mode setup is performed. The auto-detection in that case defaults to
'none' which in fact may not be supported by the underlying fs. Fix the
checker by preparing the local path in advance and creating a temporary
file for capabilities querying.

Signed-off-by: Vasiliy Ulyanov <vulyanov@suse.de>