Switch to the new Go bindings for libvirt
The new bindings (libvirt.org/go/libvirt) are, from a functional
standpoint, the same as the old ones (libvirt.org/libvirt-go);
however, along with the new import path also come the following
improvements:

  * Go modules support;

  * semantic versioning;

  * import path actually matching the package name.

The old bindings are considered obsolete and are not going to
see any further development going forward.

Since the first available release of the new bindings is
v1.7005.0 (libvirt 7.5.0), we are updating the code slightly
past what is currently used by KubeVirt (libvirt 7.3.0): this
is perfectly safe to do, because the libvirt Go bindings are
designed to work with any older version of the C library.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>