Add workaround for the issue with CPU manager
CPU manager periodically updates cgroup settings via the container
runtime interface. Runc prior to version v1.0.0 drops all 'custom'
cgroup device rules on 'update' and that causes a race with live
migration when block volumes are hotplugged. Try to setup the test in a
way so that the VMI is migrated to a node without CPU manager.

Signed-off-by: Vasiliy Ulyanov <vulyanov@suse.de>