masquerade: Calculate dhcp info on phase2
As part of this change dhcpConfigurator is now used only in phase2.
The only thing dhcpConfigurator did on phase1 is a simple export.
It isn't worth the confusion and complication of the code to maintain
dhcpConfigurator in both of the phases.

Note: the commit takes into account the other bindings would also reduce
their dhcp cache usage and puts the foundation for it.


Signed-off-by: Alona Kaplan <alkaplan@redhat.com>