Fix DHCP server test false positive
One of the e2e tests that check if virt-launcher panics in case the DHCP
server cannot startup presents a false possitive, in case the DHCP
server is not failing it will still panic so the tests will pass. The
file that is used to check if DHCP server is already started need to be
replaced with a stub since current path is not writable at unit test
time.

Also add a phase2 positive test now that the false positive is fix.

Signed-off-by: Quique Llorente <ellorent@redhat.com>