net phase1: prevent run twice with state
Currently there are scenarios where the virt-launcher pod networking setup step
can be executed multiple times, this is prevented by checking if the libvirt
domain interface cache file is already stored since this happends after the
networking setup step, this create a dependency between checking if the
networking setup step has being executed and if the libvirt domain interfaces
has being generated. To break this dependency the fact that the networking
setup step has being done is reprenseted as a "state" field at pod interface
yaml with the following values:

* PodIfaceNetworkPreparationPending: the virt-launcher networking setup has
  not being executed yet
* PodIfaceNetworkPreparationFinished: the virt-launcher networking setup has
  being successfully executed.
* PodIfaceNetworkPreparationStarted: the virt-launcher networking setup was
  interrupted before finished so networking state is unknown.

In case of executing the networking setup but the state is
PodIfaceNetworkPreparationStarted a error is returned since this step cannot
be resumed, in case the state is PodIfaceNetworkPreparationPending all
the phase1 can be executed again
(dhcp config, pod IPs and libvirt domain interface can be executed again),
if state is PodIfaceNetworkPreparationFinished nothing is done.

Signed-off-by: Quique Llorente <ellorent@redhat.com>