Add filesystem wrapper to add path prefix
To do unit testing the real path cannot be used, using the same approach
at [1] this change add a filesystem wrapper with an interface to be able
to add a prefix at all the path used so filesystem can be used at unit
test.

Also it modified a [1] to include "WriteFile" function.

[1] https://github.com/kubernetes/kubernetes/tree/master/pkg/util/filesystem

Signed-off-by: Quique Llorente <ellorent@redhat.com>