Remove fake interface cache factory
The fake InterfaceCacheFactory has reimplementation of some part of the
production version, for example how we handle the error codes if
something is not found. This change increment the use of production code
by using the real implementation but adding a base path to the
filesystem so unit tests can read/write and are isolated between each
other.

Signed-off-by: Quique Llorente <ellorent@redhat.com>