virt-launcher, hostdevice: Refactor GPU & Generic HostDevice
Refactor the GPU and generic HostDevice implementation and unify it with
the SR-IOV one.

As part of the change, the following main points have been handled:
- Generic, GPU & SR-IOV host-devices now share the same env parsing in
the form of address pools.
- The creation of PCI devices are shared between all.
- The creation of domain MDEV based devices are shared between GPU and generic
host-devices.

In terms of code-tree structure:
device-->hostdevice+->generic
                   |->gpu
                   -->sriov

The hostdevice package contains the common infrastructure.

As part of this change, all host-devices now include an alias which is
composed from a type name prefix (hostdevice, gpu, sriov), followed by
the device name (e.g. `gpu-mydev0`).

Signed-off-by: Edward Haas <edwardh@redhat.com>