Make docs/metrix.md fully auto-generated
Fixes 5834

Currently, the [docs/metrix.md](docs/metrix.md) file is partially auto-generated and partially manually edited. That makes it hard to write an automation to maintain this file, and causes confusing output.

In addition, the file currently contains the HELP line and this is also confusing, because these lines contains both the name and the description, while they are already included in the document. In some cases the description in the HELP line, is not the same as the description in the description line, and that adds to the confusion. Also, the HELP line is with heading, and that makes it not readable if it's too long.

This PR makes the help lines in the code to be the source of truth. The doc-generator tool takes the information from the code and use it to write the metrics in the document. The doc-generator tool now ignores the existing document and generates a new documenet without reading the existing one. In most cases, is identical to the existing document).

The meaning of this change is that any manual change to the docs/metrix.md document will be overwritten by the doc-generator tool.

In a few places, where the description in the HELP line was not the same as the description in the metric source code, the PR fixed the source code to match the document, so from now on, this will be the description for this metric.

And finally, the new document format is without the confusing HELP line.

Signed-off-by: Nahshon Unna-Tsameret <nunnatsa@redhat.com>