fix racy check in test 5257
The test might fail to update the vm spec because vm spec
cant be updated during vmsnapshot. The BeforeTest creates the snapshot and waits
for it to be ready and then updates the vm spec to running, but the create snapshot
function doesnt wait for the SnapshotInProgress in the vm status to be
changed back to nil and so the vm update fails.

Signed-off-by: Shelly Kagan <skagan@redhat.com>