sandbox: Add a custom sandbox to bazel
In order to allow consistent builds independent of the base operating
system, this PR introduces sandbox targets to bazel which solely have
the purpose to build a common chroot. Once the chroot is created, a
`sandbox.bazelrc` file is created which will ensure that bazel will from
now on use the content of that image as a base.

This decouples bazel completely from the base system, including glibc
installed on the system. As a consequence the builder container still
serves as a common base to ship generator binaries and bazel itself, but
bazel will now create reproducible results inside and  outside of the
build container and the build container is only necessary now for `make
generate`.

Signed-off-by: Roman Mohr <rmohr@redhat.com>