tests, Fix Relogin of fedora
LoginToFedora had a bug that in case of re-login,
where the previous command failed, but its return code
was not consumed yet, it will fail the loggedin check
and execute almost all the login flow again,
(for example reconfigure the console).

Fix it by check if the full prompt is met.
Prompt might be the one of the fedora user or the one
of root, if sudo su was executed.

Signed-off-by: Or Shoval <oshoval@redhat.com>