dockerized: make sure that docker secrets volume exists
make sure that /root/.docker/secrets exists and shared with the
builder container as a volume. In the past /root/.docker
was shared as a volume.

The /root/.docker/secrets path is used by the upstream CI infra
in order to mount the secret into the bootstrap container pod.
In the long-term we will change this location in order not to abuse
the docker configuration directory.

Signed-off-by: Igor Bezukh <ibezukh@redhat.com>