Fix flaky hostpath storage tests
Fetch the latest state of the hostpath job pod before it is returned to
the caller: previously it was returned right after the creation and thus
the pod.Spec.NodeName was empty (since it was not yet scheduled). The
NodeName is needed to schedule NFS or VMI pods on the same node later
on.

Signed-off-by: Vasiliy Ulyanov <vulyanov@suse.de>