test, network: run in parallel
Networking tests in general did not run in parallel until this commit.
0be880feb837de35ae1c93de95dea79bb6076485 and the successive three commits
removed beforeAll from five networking tests, and allowed to run them in parallel.
This commit makes all networking tests to be aligned with the rest of the tests
and run in parallel, and spare sigificant time.

Tests that modified the kubevirt configurations are not allowed to run in parallel, so this
commit also made those configurations at the suit setup, making them as default, and not per test,
allowing them to run in parallel.

Signed-off-by: alonsadan <asadan@redhat.com>