Fix virtctl imageupload on default storageclass
Previously When StorageClass is not provided then Virtctl requires the
clients to have permissions to list storageclasses and get cdi storage
profiles within the cluster, which limits who can upload images. This
is only used to get error message faster, and to provide default value.

Now the logic to get storageclasses and storageprofile was removed.

Signed-off-by: Bartosz Rybacki <brybacki@redhat.com>