tests: Make sure libguestfs appliance is fully unpacked
The guestfs tests have been checking if the appliance is ready by
searching for one of the files that is unpacked at container startup.
This isn't robust enough since the tar process could still be unpacking
the archive.  To make this check more robust, check if all expected
files are present _and_ that tar is not still running in the container.

Signed-off-by: Adam Litke <alitke@redhat.com>