Remove unsafe migration test
The only remaining case where unsafe migrations has an effect is when we
use a unsafe filesystem on a network exposed resource (e.g. ext4 on
iscsi).

In the past some cache modes were considered unsafe too. This is however
no longer the case.

Since this is the only test which we can't migrate away from iscsi and
giving the reduced importance, removing the tests to get rid of iscsi in
our test suite.

Signed-off-by: Roman Mohr <rmohr@redhat.com>