sriov, Fix flaky context deadline exceeded tests
Lately SR-IOV lane became flaky due to the warning:

"unknown error encountered sending command SyncVMI:
rpc error: code = DeadlineExceeded desc = context deadline exceeded"

This is a warning level event, which has a built-in
retry mechanism in kubevirt.

SR-IOV tests are used to ignore this warning,
but a new function, StartVmOnNode, didn't.
Therefore remove waiting for the VMI start from StartVmOnNode function.
Each caller of StartVmOnNode already waiting for start anyhow,
it was redundant, so it is safe to remove it.
Moreover it didn't ignore this specific warning, which in turn
caused this warning to fail the tests.

Potential follow-up work can research and optimize
the number of calls to syncVMI, which will reduce the
compute needed for this type of messages.

Signed-off-by: Or Shoval <oshoval@redhat.com>