macvtap, tests: Increase macvtap iface IP address report timeout
QEMU guest-agent (GA) learn and reports the network interfaces
stats of the guest.
We see it timeout's to report the macvtap interface IP address
for a while now, it may take a bit more time for the GA to
report it.
As initial step we should increase the time we wait for the IP
address to appear.

Signed-off-by: Or Mergi <ormergi@redhat.com>