Restructure core/v1 to be compatible with client-gen
The kubecli client has special handling of v1alpha3 and v1. It allows
for instance using an env var to not register all versions. This logic
is move to kubecli.go and out of apis/core to avoid conflicts with
client-gen.

The final intent is that client-gen can focus on a clean v1 while
client-go still supports the whole set of workarounds which some other
projects which vendor client-go rely on.

Signed-off-by: Roman Mohr <rmohr@redhat.com>