virctl: Fix parameters default value
The value of some parameters was demanded to the value of the variable where it will be stored. In works in production but it does not in unit tests.
So if a test for example uses --force for restart command, the next test continue to have the value of the variable force as true.

Signed-off-by: fossedihelm <fossedihelm@gmail.com>