rpm: Pin QEMU version everywhere
Currently testimage will end up with some random version of
qemu-img: probably the latest one, but possibly not the same
one as launcherbase.

This will not necessarily cause any problem, but it's nicer
to use the same version everywhere. It also saves on download
time during build and disk usage for the artifact mirror.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>