tests, network, istio: revert parallelization
The istio tests create istio specific resources that affects
the istio-proxy container configuration in all namespaces. When
running in parallel, the tests can flake.

This commit reverts the changes made in 5936e7c9e1f8366f351ec278e9fda00f757ef82c
and make the istio tests run in serial. In the future commit the tests
should be implemented in a way that allows them to be run in parallel.

Signed-off-by: alonsadan <asadan@redhat.com>