virt-handler, net: Drop virt-launcher pid usage in cache
In order to avoid network setup on phase 1 to be executed more than
once, a cache check exists. This "cache" uses the VMI UID as the key and
the virt-launcher PID as the value.

There is no apparent reason to include the PID and validate it did not
changed over time. If the virt-launcher terminates, a full teardown is
expected to occur, including this cache cleanup for the specific VMI.

Therefore, this change removes the PID usage in the cache.

Signed-off-by: Edward Haas <edwardh@redhat.com>