istio, network: forward port 22 for istio aware VMIs
Istio by default excludes port 22 from proxying. The reason for this
exclusion is for use cases where istio proxies run on VMs, to prevent
loosing connectivity to nodes in case of misconfiguration of the proxies.

In our use case however, we're forwarding to the VM network namespace
only traffic that was proxied. Therefore, port 22 is missed as it's not
proxied.

Current implementation of this port 22 exclusion in Istio u/s doesn't
give us other option than add an exception for this port to make sure that
traffic to port 22 is forwarded to the VM.

Note: This means that traffic to 22 is not being proxied, but the same behaviour
would be with regular pods.

This is the line of code in Istio u/s that adds the port 22 exclusion:
https://github.com/istio/istio/blob/f907fd498846a877fab3c1013ffdb6cd83bec28a/tools/istio-iptables/pkg/capture/run.go#L172

The only way around ATM would be adding an anotation that explicitly
defines list of ports to proxy, but that's not a viable option
for use case where we forward all ports.

We may revert this when https://issues.redhat.com/browse/OSSM-782
is resolved.

Signed-off-by: Radim Hrazdil <rhrazdil@redhat.com>