Add configured capabilites according to storage
Added to existing config files and add a new config
ceph file which configures the storageclass that can be used
when needed specific need and not specific storage.
For example tests which required OCS when actually just need
RWX Block storage. Now they will be able to ask for RWX block
storage class which can be changes according to the lane,
hence the test wont be confined to run with specific storage
class and will be able to be tested with several storages.

Signed-off-by: Shelly Kagan <skagan@redhat.com>