Replace hard coded use of OCS with sc matching with test requirements
Some tests required OCS storage class because of its capabilities
regardless if there might be other storage class available which
similat capabilities. Those tests were genrealized to ask for a
storage class for the actual required capability, for example
for RWO or RWX block storage.
In those places changed the [rook-ceph] tag to [block-storage] tag.
which currently will run in the sig-storage lane or lane with
ceph storage.

Signed-off-by: Shelly Kagan <skagan@redhat.com>