netstat: Remove test that assumes domain spec without vmi spec
In the current implementation, the domain spec is directly affected by
the VMI spec (in the network interfaces scope).
The scenario were the domain is configured with interfaces that do not
appear in the VMI spec is no supported (or expected).

Therefore, the test that checks such a scenario (in an implicit manner)
is removed.

Signed-off-by: Edward Haas <edwardh@redhat.com>