netstat: Add misc unit tests for undefined scenarios
The unit tests included in this change represent undefined behavior.
These scenarios are explicitly represented through dedicated tests to
show the expected result if such a case occurs.

Future changes may suggest to change this output and replace it with
something that fits better (e.g. do not show the interface entry).

Signed-off-by: Edward Haas <edwardh@redhat.com>