Fix misleading domain memory metrics
kubevirt_vmi_memory_used_total_bytes refers the amount of
memory declared in libvirt domain xml file. It is misleading.
We decided to rename it as "kubevirt_vmi_memory_domain_total_bytes".

We also think it is valuable to have a metric which gives
the amount of memory used in the VM. We define a new metric
"kubevirt_vmi_memory_used_bytes" for this purpose. It is computed as
"kubevirt_vmi_memory_available_bytes-kubevirt_vmi_memory_usable_bytes"

For all details, See  https://bugzilla.redhat.com/show_bug.cgi?id=2018925

Signed-off-by: Erkan Erol <eerol@redhat.com>