api: Add UnmarshalJSON method to PodNetwork
Added spacial handling for the VMNetworkCIDR field.
In order to make it backwards compatible on go 1.17,
we remove leading zeros from CIDR notations,
so the rest of the system won't have spacial treatment for it.

This also adds a function to remove leading zeros from IPv4 CIDR
notations.
This function is protocol agnostic and can handle IPv6 CIDR notations
(return them as-is).

Signed-off-by: Orel Misan <omisan@redhat.com>