network, netstat: Simplify status update
This change aims to simplify the status update logic for network
interface.

It follows the expected logic where:
- The interface status is populated with IP data from the configuration
cache (written to during the network setup).
- Afterwards, it gets updated by the data collected from the domain
configuration.
- Lastly, it gets updated by the (optional) guest-agent data.

With the new logic in place, the SR-IOV reporting is partially fixed to
not report interfaces when the guest-agent is missing.
A full fix should come in a following change that resolves
https://github.com/kubevirt/kubevirt/issues/7050

Signed-off-by: Edward Haas <edwardh@redhat.com>